N325-036 Flatten series of Proposition separated by And
Instead of having one hypothesis with multiple propositions
separated by and operators implying a term, we want to
separate them into a serie of multiple hypothesis implying
the same term as before.

* src/tranform/eliminate_unused_hypo.ml
(flatten_conjunction): This is the function that flattens
hypothesis containing multiple conjunctions.

(descend_to_goal): added call to flatten_conjunction at the
level of an implication node as we descend into the tree