N325-036 eliminate_unused_hypo improved
Added the simplification of unused terms inside hypothesis that are
conjunction of disjunctions.

* src/transform/eliminate_unused_hypo.ml
(part_needed_hypothesis): Added recursivity to handle the cascading effect
of putting back hypothesis and added the computation of a map of the new
used variable.
(hypo_deps): New function, computes the variables on which the
removal/simplification of a hypothesis will depend.
(simpl_hypo): New function, simplifies a hypothesis when we find a
quantifier which introduces a variable on which a hypothesis depends
(clear_hypothesis_on_quant): added code to put hypothesis back if they
can't be simplified any more and return a new term with added hypothesis