This commits adds the use of Session_itp in place of old Session.
The main changes to make the testsuite pass are the following:
- proof_state are now included in the controller_itp of Why3 together with
the provers. These things dont need to be computed, they can be queried
directly from Why3.
- Gnat_scheduler has been added. It is a scheduler specifically intended
for gnatwhy3. It is based on Unix_scheduler with input string removed and
a new function to compute transformations before continuing (which is used
to keep the old behaviour).
The first tests of testsuite are passing. Expected regressions on manual
proofs.

* Makefile.in
(GNAT_FILES): Adding gnat_scheduler

* src/gnat/gnat_config.ml
(provers, prover_ce, config, env): In this function, we change directly
the gnat_config and save it instead of having functions that prepend the
drivers directory everywhere. Note that this is used by the controller and
one dont need to look at it after controller is initialized.
Also, we don't load drivers at this moment. We keep only the provers (not
their config which can be easily queried in the controller).

(build_shared): The type of the prover argument changes. It changes
accordingly everywhere.

(manual_prover): Rewriting using Whyconf functions instead of something
stored in Session.

* src/gnat/gnat_objectives.ml
(type key): Key is not needed in new sessions. Nodes of the proof tree
already have an id (see Session_itp.ml and Controller_itp.ml).

(goal -> goal_id): with the goal_id and the session we can retrieve all
informations that were formerly contained in type goal. This leads to some
changes: adding sessions/controller to a lot of elementary functions. This
is unavoidable and this is how Why3 API is working now.

(my_session): the session is now contained in what we call a controller
which is used for all actions on the sessions (schedule transformations/
prover etc).

(GoalHash): we can check equality on id of goals in the same way.

(GoalCmp): Using goal_id for equality on goals.

(GoalMap): Now an alias for proof_node hashtable from session.

(init_cont): Adding a function to initialize session and controller.

(all): Removed unused functions.

* src/gnat/gnat_manual.ml
(all): Did not care about retrieving manual proofs yet.

* src/gnat/gnat_scheduler.ml
Adding an initial scheduler inspired from Unix_scheduler without the stdin
input and with an added wait_for_idle to recover the old behaviour of
running transformations directly after scheduling them.

* src/gnat/gnat_main.ml
(): Adapted things. Some stuff still needs adaptation like manual proofs,
displaying of results and the way we init and end scheduling.


Change-Id: I437cf2cf4b0ec9cd3f5e7bb4dbe3365f2447fa34